Add Absencebook function to Attendance ManagMonthUser
